James branch (#10)
* Made changes to the backend to make it more immutable and minimize error risk when using the enum. Changed Tile.java and Piece.java to abstract object classes that each child class would extend. This allows each indivdual piece to Override small details that pertain to its individual responsibility while minimizing coding changed.
Board.java was also changed in order to make the build a little more abstract along with allowing 2 points of safety when it comes to error control. Tiles are FINAL therefore cannot move same with piece data. This allows the board to remain the same and each piece to simply move around. If you have any question please let me know.

* Made first draft of dice roll implementation with dice animation. Needs to be switched from Box to a TriangleMesh and MeshView in order for the sides to each be different. This will be a task for a later date.

* Working on Completing moves:
BISHOP and PAWN are complete.
Rest will throw error currently as they are null.

* KING, QUEEN, and KNIGHT are working however needs some work. The recursive break needs to be more finite to keep it from adding to many allowed moves.

Co-authored-by: Damen <damendeberry@aol.com>